ai-powered-hybrid-search/requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.34.0
torch==2.0.1
numpy==1.24.3
scipy==1.10.1
elasticsearch==8.6.1
fastapi==0.103.1
uvicorn==0.22.0